Fix import bug
