+++ models/cnn/famous_cnn.py
@@ -26,26 +26,3 @@ architectures  = {
     "vgg19": vgg19.vgg19
     }
 
-def correct_pad(inputs, kernel_size):
-    """Returns a tuple for zero-padding for 2D convolution with downsampling.
-    # Arguments
-        input_size: An integer or tuple/list of 2 integers.
-        kernel_size: An integer or tuple/list of 2 integers.
-    # Returns
-        A tuple.
-    """
-    img_dim = 2 if keras.backend.image_data_format() == 'channels_first' else 1
-    input_size = keras.backend.int_shape(inputs)[img_dim:(img_dim + 2)]
-
-    if isinstance(kernel_size, int):
-        kernel_size = (kernel_size, kernel_size)
-
-    if input_size[0] is None:
-        adjust = (1, 1)
-    else:
-        adjust = (1 - input_size[0] % 2, 1 - input_size[1] % 2)
-
-    correct = (kernel_size[0] // 2, kernel_size[1] // 2)
-
-    return ((correct[0] - adjust[0], correct[0]),
-            (correct[1] - adjust[1], correct[1]))
