+++ models/cnn/mobilenetv2.py
@@ -9,8 +9,31 @@ This file contains building code for MobileNetV2, based on
 """
 
 import tensorflow.keras as keras
-from . import famous_cnn.correct_pad as correct_pad
 
+
+def correct_pad(inputs, kernel_size):
+    """Returns a tuple for zero-padding for 2D convolution with downsampling.
+    # Arguments
+        input_size: An integer or tuple/list of 2 integers.
+        kernel_size: An integer or tuple/list of 2 integers.
+    # Returns
+        A tuple.
+    """
+    img_dim = 2 if keras.backend.image_data_format() == 'channels_first' else 1
+    input_size = keras.backend.int_shape(inputs)[img_dim:(img_dim + 2)]
+
+    if isinstance(kernel_size, int):
+        kernel_size = (kernel_size, kernel_size)
+
+    if input_size[0] is None:
+        adjust = (1, 1)
+    else:
+        adjust = (1 - input_size[0] % 2, 1 - input_size[1] % 2)
+
+    correct = (kernel_size[0] // 2, kernel_size[1] // 2)
+
+    return ((correct[0] - adjust[0], correct[0]),
+            (correct[1] - adjust[1], correct[1]))
 
 def _make_divisible(v, divisor, min_value=None):
     if min_value is None:
@@ -53,7 +76,7 @@ def _inverted_res_block(inputs,
         prefix = 'expanded_conv_'
 
     if stride == 2:
-        x = keras.layers.ZeroPadding2D(padding=correct_pad(keras.backend, x, 3),
+        x = keras.layers.ZeroPadding2D(padding=correct_pad(x, 3),
                                  name=prefix + 'pad')(x)
     x = keras.layers.DepthwiseConv2D(kernel_size=3,
                                strides=stride,
